feat: change data source to lxns for wordle

+++ util/Data.py
@@ -13,7 +13,7 @@ alias_list_ycn_lock = Lock()
 alias_list_xray_lock = Lock()
 
 
-async def get_music_data():
+async def get_music_data_df():
     cache_dir = "./Cache/Data/MusicData/"
     cache_path = f"{cache_dir}{date.today().isoformat()}.json"
     async with music_data_lock:
@@ -23,6 +23,34 @@ async def get_music_data():
                 try:
                     async with session.get(
                         "https://www.diving-fish.com/api/maimaidxprober/music_data"
+                    ) as resp:
+                        async with aiofiles.open(cache_path, "wb") as fd:
+                            await fd.write(await resp.read())
+                except ClientError:
+                    if os.path.exists(cache_path):
+                        os.remove(cache_path)
+                    if files:
+                        async with aiofiles.open(f"{cache_dir}{files[-1]}") as fd:
+                            return json.loads(await fd.read())
+                    return list()
+            if files:
+                for file in files:
+                    os.remove(f"{cache_dir}{file}")
+    async with aiofiles.open(cache_path) as fd:
+        return json.loads(await fd.read())
+
+
+async def get_music_data_lxns():
+    cache_dir = "./Cache/Data/MusicDataLxns/"
+    cache_path = f"{cache_dir}{date.today().isoformat()}.json"
+    async with music_data_lock:
+        if not os.path.exists(cache_path):
+            files = os.listdir(cache_dir)
+            async with ClientSession(conn_timeout=3) as session:
+                try:
+                    async with session.get(
+                        "https://maimai.lxns.net/api/v0/maimai/song/list",
+                        params={"notes": "true"},
                     ) as resp:
                         async with aiofiles.open(cache_path, "wb") as fd:
                             await fd.write(await resp.read())
